Update qaqc and director features
