+++ app/Http/Controllers/qaqc/QaqcHomeController.php
@@ -1,21 +1,19 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\qaqc;
 
 use App\Http\Controllers\Controller;
 use App\Models\Report;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 
-class StaffHomeController extends Controller
+class QaqcHomeController extends Controller
 {
     public function index()
     {
+
         $approvedDoc = Report::where('is_approve', 1)->count();
         $waitingDoc = Report::whereNull('is_approve')->count();
         $rejectedDoc = Report::where('is_approve', 0)->count();
 
-        return view('staff_home', compact('approvedDoc', 'waitingDoc','rejectedDoc'));
+        return view('qaqc.home', compact('approvedDoc', 'waitingDoc','rejectedDoc'));
     }
-
 }
