+++ resources/views/director/qaqc/index.blade.php
@@ -2,14 +2,24 @@
 
 @section('content')
 <section class="header">
-    <div class="d-flex mb-3 row-flex">
-        <div class="h2 p-2 me-auto">QA/QC Reports</div>
+    <div class="d-flex mb-1 row-flex">
+        <div class="h2 me-auto">QA & QC Reports</div>
     </div>
 </section>
 
+<section>
+    <nav aria-label="breadcrumb">
+        <ol class="breadcrumb mb-5 ">
+            <li class="breadcrumb-item"><a href="{{route('director.home')}}">Dashboard</a></li>
+            <li class="breadcrumb-item active" aria-current="page">QA & QC Reports</li>
+        </ol>
+    </nav>
+</section>
+
 @if ($message = Session::get('success'))
-    <div class="alert alert-success">
-        <p>{{ $message }}</p>
+    <div class="alert alert-success alert-dismissible fade show" role="alert">
+        {{ $message }}
+        <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
     </div>
 @endif
 
@@ -19,7 +29,7 @@
             <table class="table table-hover table-bordered mb-0 text-center table-striped">
                 <thead>
                     <tr>
-                      <th class="fs-5" scope="col">ID</th>
+                      <th class="fs-5" scope="col">No</th>
                       <th class="fs-5" scope="col">Invoice No</th>
                       <th class="fs-5" scope="col">Customer</th>
                       <th class="fs-5" scope="col">Verify Date</th>
@@ -32,20 +42,22 @@
                     @foreach ($reports as $report)
                         @if ($report->autograph_1 && $report->autograph_2 && $report->autograph_3)
                         <tr>
-                            <td>{{ $report->id }}</td>
+                            <td>{{ $loop->iteration }}</td>
                             <td>{{ $report->invoice_no }}</td>
                             <td>{{ $report->customer }}</td>
                             <td>{{ $report->rec_date }}</td>
                             <td>{{ $report->verify_date }}</td>
-                            <td class="text-start">
-                                <a href="{{ route('direktur.qaqc.detail', ['id' => $report->id]) }}" class="btn btn-info" style="--bs-bg-opacity: .5;">View Details</a>
+                            <td>
+                                <a href="{{ route('director.qaqc.detail', ['id' => $report->id]) }}" class="btn btn-secondary">
+                                    <i class='bx bx-info-circle' ></i> Detail
+                                </a>
                                 @if($report->attachment)
                                     @php
                                         $filename = basename($report->attachment);
                                     @endphp
-                                    <a href="{{ asset('storage/attachments/' . $report->attachment) }}" class="btn btn-info" style="--bs-bg-opacity: .5;" download="{{ $filename }}">
+                                    <a href="{{ asset('storage/attachments/' . $report->attachment) }}" class="btn btn-success" download="{{ $filename }}">
                                         <i class='bx bx-download'></i>
-                                        Download Attachment
+                                        Attachment
                                     </a>
                                 @endif
                             </td>
@@ -54,7 +66,7 @@
                                     <span class="badge rounded-pill text-bg-success px-3 py-2 fs-6 fw-medium">APPROVED</span>
                                 @elseif($report->is_approve === 0)
                                     <span class="badge rounded-pill text-bg-danger px-3 py-2 fs-6 fw-medium">REJECTED</span>
-                                    @else
+                                @else
                                     <span class="badge rounded-pill text-bg-warning px-3 py-2 fs-6 fw-medium">WAITING</span>
                                 @endif
                             </td>
